Add loading button
